+++ packages/web/menu/src/menu.tsx
@@ -1,5 +1,6 @@
-import React, { ReactNode } from 'react';
+import React, { ReactNode, useCallback, FC } from 'react';
 import { IMenuProps, MenuMain } from './menuMain';
+import { Popover } from '@elonwu/web-popover';
 
 interface IMenuDataProps {
   key?: string;
@@ -13,42 +14,59 @@ interface IMenuDataProps {
 
 export interface IMenuCMPProps extends IMenuProps {
   data: IMenuDataProps[];
+  popover?: boolean;
 }
 
-export const Menu = (props: IMenuCMPProps) => {
-  const { data } = props;
+export const Menu: FC<IMenuCMPProps> = (props) => {
+  const { data, popover } = props;
 
-  const renderMenu = (oriData: IMenuDataProps[]) => {
-    const node = oriData.map((item) => {
-      const {
-        key,
-        name,
-        icon,
-        disabled,
-        subMenus,
-        showSubMenus,
-        render,
-      } = item;
-      if (subMenus && subMenus.length > 0) {
-        return (
-          <MenuMain.SubMenu key={key} title={name} showSubMenus={showSubMenus}>
-            {renderMenu(subMenus)}
-          </MenuMain.SubMenu>
-        );
-      } else {
-        return (
-          <MenuMain.Item
-            key={key}
-            disabled={disabled}
-            icon={render ? '' : icon}
-          >
-            {render ? render({ ...item }) : name}
-          </MenuMain.Item>
-        );
-      }
-    });
-    return node;
-  };
+  const renderMenu = useCallback(
+    (oriData: IMenuDataProps[]) => {
+      const node = oriData.map((item) => {
+        const {
+          key,
+          name,
+          icon,
+          disabled,
+          subMenus,
+          showSubMenus,
+          render,
+        } = item;
+        if (subMenus && subMenus.length > 0) {
+          return popover ? (
+            <Popover
+              position="rightBottom"
+              trigger={<MenuMain.SubMenu key={key} title={name} />}
+              style={{ width: '100%' }}
+              contentStyle={{ width: 120 }}
+            >
+              {renderMenu(subMenus)}
+            </Popover>
+          ) : (
+            <MenuMain.SubMenu
+              key={key}
+              title={name}
+              showSubMenus={showSubMenus}
+            >
+              {renderMenu(subMenus)}
+            </MenuMain.SubMenu>
+          );
+        } else {
+          return (
+            <MenuMain.Item
+              key={key}
+              disabled={disabled}
+              icon={render ? '' : icon}
+            >
+              {render ? render({ ...item }) : name}
+            </MenuMain.Item>
+          );
+        }
+      });
+      return node;
+    },
+    [popover],
+  );
 
   return (
     <div>