+++ packages/web/portal/src/index.ts
@@ -1,4 +1,3 @@
 export * from './modal';
 export * from './drawer';
 export * from './message';
-export * from './popover';